+++ Engine/Plugins/Runtime/HairStrands/Source/HairStrandsCore/Private/HairStrandsRendering.cpp
@@ -66,7 +66,7 @@ bool NeedsUpdateCardsMeshTriangles()
 static FIntVector ComputeDispatchCount(uint32 ItemCount, uint32 GroupSize)
 {
 	const uint32 BatchCount = FMath::DivideAndRoundUp(ItemCount, GroupSize);
-	const uint32 DispatchCountX = FMath::FloorToInt(FMath::Sqrt(BatchCount));
+	const uint32 DispatchCountX = FMath::FloorToInt(FMath::Sqrt(static_cast<float>(BatchCount)));
 	const uint32 DispatchCountY = DispatchCountX + FMath::DivideAndRoundUp(BatchCount - DispatchCountX * DispatchCountX, DispatchCountX);
 
 	check(DispatchCountX <= 65535);
@@ -78,7 +78,7 @@ static FIntVector ComputeDispatchCount(uint32 ItemCount, uint32 GroupSize)
 // Same as above but the group count is what matters and is preserved
 static FIntVector ComputeDispatchGroupCount2D(uint32 GroupCount)
 {
-	const uint32 DispatchCountX = FMath::FloorToInt(FMath::Sqrt(GroupCount));
+	const uint32 DispatchCountX = FMath::FloorToInt(FMath::Sqrt(static_cast<float>(GroupCount)));
 	const uint32 DispatchCountY = DispatchCountX + FMath::DivideAndRoundUp(GroupCount - DispatchCountX * DispatchCountX, DispatchCountX);
 
 	check(DispatchCountX <= 65535);
